backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pydantic==2.10.4
python-dotenv==1.0.1
certifi==2024.12.14
twilio>=8.0.0
websockets==14.1
